move TeeReader to lib
