+++ chapter3/src/3_7/tee_reader.rs
@@ -1,35 +1,5 @@
-use std::io::{Read, Write};
-
-struct TeeReader<R, W>
-where
-    R: Read,
-    W: Write,
-{
-    reader: R,
-    writer: W,
-}
-
-impl<R, W> TeeReader<R, W>
-where
-    R: Read,
-    W: Write,
-{
-    fn new(reader: R, writer: W) -> Self {
-        Self { reader, writer }
-    }
-}
-
-impl<R, W> Read for TeeReader<R, W>
-where
-    R: Read,
-    W: Write,
-{
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        let n = self.reader.read(buf)?;
-        self.writer.write_all(&buf[..n])?;
-        Ok(n)
-    }
-}
+use lib::io::TeeReader;
+use std::io::Read;
 
 fn main() -> std::io::Result<()> {
     // TeeReader を使用した場合の例。
