+++ lib/src/io.rs
@@ -2,8 +2,10 @@ mod limited_reader;
 mod multi_reader;
 mod multi_writer;
 mod section_reader;
+mod tee_reader;
 
 pub use limited_reader::LimitedReader;
 pub use multi_reader::MultiReader;
 pub use multi_writer::MultiWriter;
 pub use section_reader::SectionReader;
+pub use tee_reader::TeeReader;
